Add stability precision setting
This corresponds roughly with how still you have to stand on the balance
board for your weighin to occur. The lower it is, the more you can move
around but the less accurate the weight may be. The higher it is, the
less you can move around but the more accurate the weight may be.